Change type of tokens to BigInteger

+++ src/TokenSystem.Tests/TestFungibleTaggingAndPicking.cs
@@ -1,3 +1,4 @@
+using System.Numerics;
 using ContractsCore;
 using TokenSystem.Tokens;
 using TokenSystem.Exceptions;
@@ -14,7 +15,7 @@ namespace TokenSystem.Tests
 
 		[Theory]
 		[InlineData(100)]
-		public void Tag_WhenPassedValidAddressAndAmount_TagsTokensEvenly(decimal amount)
+		public void Tag_WhenPassedValidAddressAndAmount_TagsTokensEvenly(BigInteger amount)
 		{
 			IReadOnlyTaggedTokens<string> tokens = this.fungibleTagger.Tag(this.defaultAddress, amount);
 			Assert.Equal(amount, tokens.GetAmountByTag(FungibleTokenTagger.DefaultTokenTag));
@@ -22,7 +23,7 @@ namespace TokenSystem.Tests
 
 		[Theory]
 		[InlineData(-100)]
-		public void Tag_WhenPassedNegativeAmount_Throws(decimal amount)
+		public void Tag_WhenPassedNegativeAmount_Throws(BigInteger amount)
 		{
 			Assert.Throws<NonPositiveTokenAmountException>(
 				() => this.fungibleTagger.Tag(this.defaultAddress, amount));
@@ -30,7 +31,7 @@ namespace TokenSystem.Tests
 
 		[Theory]
 		[InlineData(23, 3)]
-		public void Pick_WhenRequestedLessThanOwnedTokens_ReturnsCorrectAmount(decimal mintAmount, decimal pickAmount)
+		public void Pick_WhenRequestedLessThanOwnedTokens_ReturnsCorrectAmount(BigInteger mintAmount, BigInteger pickAmount)
 		{
 			IReadOnlyTaggedTokens<string> tokens = this.fungibleTagger.Tag(this.defaultAddress, mintAmount);
 			IReadOnlyTaggedTokens<string> pickedTokens = this.fungiblePicker.Pick(tokens, pickAmount);
@@ -40,8 +41,8 @@ namespace TokenSystem.Tests
 		[Theory]
 		[InlineData(23, 100)]
 		public void Pick_WhenRequestedMoreThanOwnedTokens_Throws(
-			decimal mintAmount,
-			decimal pickAmount)
+			BigInteger mintAmount,
+			BigInteger pickAmount)
 		{
 			IReadOnlyTaggedTokens<string> tokens = this.fungibleTagger.Tag(this.defaultAddress, mintAmount);
 			Assert.Throws<InsufficientTokenAmountException>(() => this.fungiblePicker.Pick(tokens, pickAmount));
@@ -51,8 +52,8 @@ namespace TokenSystem.Tests
 		[Theory]
 		[InlineData(100, -10)]
 		public void Pick_WhenPassedNegativeAmount_Throws(
-			decimal mintAmount,
-			decimal pickAmount)
+			BigInteger mintAmount,
+			BigInteger pickAmount)
 		{
 			IReadOnlyTaggedTokens<string> tokens = this.fungibleTagger.Tag(this.defaultAddress, mintAmount);
 			Assert.Throws<NonPositiveTokenAmountException>(() => this.fungiblePicker.Pick(tokens, pickAmount));
