+++ src/TokenSystem/Exceptions/NonPositiveTokenAmountException.cs
@@ -1,10 +1,11 @@
 using System;
+using System.Numerics;
 
 namespace TokenSystem.Exceptions
 {
     public class NonPositiveTokenAmountException : Exception
     {
-        public NonPositiveTokenAmountException(decimal tokenAmount) : base(
+        public NonPositiveTokenAmountException(BigInteger tokenAmount) : base(
             $"Non-positive token amount requested: {tokenAmount}.")
         {
         }
