+++ src/TokenSystem/TokenEventArgs/TokensMintedEventArgs.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Numerics;
 using ContractsCore;
 using TokenSystem.Tokens;
 
@@ -6,14 +7,14 @@ namespace TokenSystem.TokenEventArgs
 {
 	public class TokensMintedEventArgs<TTagType> : EventArgs
 	{
-		public TokensMintedEventArgs(decimal amount, IReadOnlyTaggedTokens<TTagType> tokens, Address to)
+		public TokensMintedEventArgs(BigInteger amount, IReadOnlyTaggedTokens<TTagType> tokens, Address to)
 		{
 			this.Amount = amount;
 			this.Tokens = tokens;
 			this.To = to;
 		}
 
-		public decimal Amount { get; }
+		public BigInteger Amount { get; }
 
 		public IReadOnlyTaggedTokens<TTagType> Tokens { get; }
