+++ src/TokenSystem/TokenManager/Actions/TransferAction.cs
@@ -1,3 +1,4 @@
+using System.Numerics;
 using ContractsCore;
 using ContractsCore.Actions;
 
@@ -10,7 +11,7 @@ namespace TokenSystem.TokenManager.Actions
 			Address origin,
 			Address sender,
 			Address target,
-			decimal amount,
+			BigInteger amount,
 			Address from,
 			Address to,
 			ITokenPicker<TTokenTagType> customPicker = null)
@@ -22,7 +23,7 @@ namespace TokenSystem.TokenManager.Actions
 			this.CustomPicker = customPicker;
 		}
 
-		public decimal Amount { get; }
+		public BigInteger Amount { get; }
 
 		public Address From { get; }
