+++ src/TokenSystem/TokenManager/FungibleTokenPicker.cs
@@ -1,4 +1,5 @@
 using System.Collections.Generic;
+using System.Numerics;
 using TokenSystem.Exceptions;
 using TokenSystem.Tokens;
 
@@ -6,21 +7,21 @@ namespace TokenSystem.TokenManager
 {
 	public class FungibleTokenPicker : ITokenPicker<string>
 	{
-		public IReadOnlyTaggedTokens<string> Pick(IReadOnlyTaggedTokens<string> tokens, decimal amount)
+		public IReadOnlyTaggedTokens<string> Pick(IReadOnlyTaggedTokens<string> tokens, BigInteger amount)
 		{
 			if (amount < 0)
 			{
 				throw new NonPositiveTokenAmountException(amount);
 			}
 
-			decimal availableTokens = tokens.GetAmountByTag(FungibleTokenTagger.DefaultTokenTag);
+			BigInteger availableTokens = tokens.GetAmountByTag(FungibleTokenTagger.DefaultTokenTag);
 
 			if (availableTokens < amount)
 			{
 				throw new InsufficientTokenAmountException(availableTokens, amount);
 			}
 
-			var pickedTokens = new ReadOnlyTaggedTokens<string>(new SortedDictionary<string, decimal>
+			var pickedTokens = new ReadOnlyTaggedTokens<string>(new SortedDictionary<string, BigInteger>
 			{
 				[FungibleTokenTagger.DefaultTokenTag] = amount
 			});
