+++ src/TokenSystem/TokenManager/FungibleTokenTagger.cs
@@ -1,4 +1,5 @@
 using System.Collections.Generic;
+using System.Numerics;
 using ContractsCore;
 using TokenSystem.Exceptions;
 using TokenSystem.Tokens;
@@ -9,14 +10,14 @@ namespace TokenSystem.TokenManager
 	{
 		public const string DefaultTokenTag = "FungibleToken";
 
-		public IReadOnlyTaggedTokens<string> Tag(Address owner, decimal amount)
+		public IReadOnlyTaggedTokens<string> Tag(Address owner, BigInteger amount)
 		{
 			if (amount < 0)
 			{
 				throw new NonPositiveTokenAmountException(amount);
 			}
 
-			var tokens = new ReadOnlyTaggedTokens<string>(new SortedDictionary<string, decimal>
+			var tokens = new ReadOnlyTaggedTokens<string>(new SortedDictionary<string, BigInteger>
 			{
 				[DefaultTokenTag] = amount
 			});
