+++ src/TokenSystem/TokenManager/ITokenPicker.cs
@@ -1,9 +1,10 @@
+using System.Numerics;
 using TokenSystem.Tokens;
 
 namespace TokenSystem.TokenManager
 {
 	public interface ITokenPicker<TTagType>
 	{
-		IReadOnlyTaggedTokens<TTagType> Pick(IReadOnlyTaggedTokens<TTagType> tokens, decimal amount);
+		IReadOnlyTaggedTokens<TTagType> Pick(IReadOnlyTaggedTokens<TTagType> tokens, BigInteger amount);
 	}
 }
