+++ src/TokenSystem/Tokens/TaggedTokens.cs
@@ -1,10 +1,11 @@
+using System.Numerics;
 using TokenSystem.TokenManager;
 
 namespace TokenSystem.Tokens
 {
 	public class TaggedTokens<TTagType> : ReadOnlyTaggedTokens<TTagType>, ITaggedTokens<TTagType>
 	{
-		public void AddToBalance(TTagType tag, decimal amount)
+		public void AddToBalance(TTagType tag, BigInteger amount)
 		{
 			TokensUtility.RequirePositiveAmount(amount);
 			if (!this.TagsToBalances.ContainsKey(tag))
@@ -18,13 +19,13 @@ namespace TokenSystem.Tokens
 
 		public void AddToBalance(IReadOnlyTaggedTokens<TTagType> tokens)
 		{
-			foreach ((TTagType tag, decimal amount) in tokens)
+			foreach ((TTagType tag, BigInteger amount) in tokens)
 			{
 				this.AddToBalance(tag, amount);
 			}
 		}
 
-		public void RemoveFromBalance(TTagType tag, decimal amount)
+		public void RemoveFromBalance(TTagType tag, BigInteger amount)
 		{
 			TokensUtility.RequirePositiveAmount(amount);
 			if (!this.TagsToBalances.ContainsKey(tag))
@@ -38,7 +39,7 @@ namespace TokenSystem.Tokens
 
 		public void RemoveFromBalance(IReadOnlyTaggedTokens<TTagType> tokens)
 		{
-			foreach ((TTagType tag, decimal amount) in tokens)
+			foreach ((TTagType tag, BigInteger amount) in tokens)
 			{
 				this.RemoveFromBalance(tag, amount);
 			}
